added authentication
added auth, dynamic delete and hide button. hiding still not implemented

+++ src/main/java/com/anoki/simpleBoard/controller/DefaultController.java
@@ -3,6 +3,8 @@ package com.anoki.simpleBoard.controller;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,6 +15,7 @@ import org.springframework.web.servlet.ModelAndView;
 
 import com.anoki.simpleBoard.models.Post;
 import com.anoki.simpleBoard.models.Tag;
+import com.anoki.simpleBoard.models.User;
 //import com.anoki.simpleBoard.models.User;
 import com.anoki.simpleBoard.service.PostService;
 import com.anoki.simpleBoard.service.PostsTagsService;
@@ -35,13 +38,13 @@ public class DefaultController {
     private static Logger logger = LoggerFactory.getLogger(DefaultController.class); 
 	
     @GetMapping("/")
-    public String home1() {
-        return "posts";
+    public String root() {
+        return "redirect:/posts";
     }
 
     @GetMapping("/home")
     public String home() {
-        return "posts";
+        return "home";
     }
 
     @GetMapping("/admin")
@@ -66,18 +69,21 @@ public class DefaultController {
     
     @GetMapping("/posts")
 	public ModelAndView getPosts() {
+    	Authentication auth = SecurityContextHolder.getContext().getAuthentication();
     	ModelAndView posts = new ModelAndView("posts");
     	posts.addObject("postList", postService.findAll());
-    	posts.addObject("post", new Post());
+    	posts.addObject("post", new Post(userService.findByUsername(auth.getName())));
     	posts.addObject("userList", userService.findAll()); 
     	posts.addObject("tagList", tagService.findAll());
-    return posts;
+    	posts.addObject("loggedUser", userService.findByUsername(auth.getName()));
+    	return posts;
 	}
 
-    @PostMapping(path = "/updatePost", params="action=save")
-    public String submit(@ModelAttribute("post") Post post, @RequestParam("tags") String tags, BindingResult result, ModelAndView model) {
+    @PostMapping(path = "/posts", params="action=save")
+    public String save(@RequestParam("idUserNew") Integer idUser, @RequestParam("textNew") String text, @RequestParam("tags") String tags) {
         //Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        //logger.info("logged user: " + auth.getName());  	
+        //logger.info("logged user: " + auth.getName());  
+    	Post post = new Post(userService.findByIdUser(idUser), text);
     	postService.addPost(post);
     	logger.info("tags passati: " + tags);
     	postsTagsService.mergeByName(tags);
@@ -85,25 +91,46 @@ public class DefaultController {
     	return "redirect:/posts";
     }
     
-    @PostMapping(path = "/deletePost")
+    @PostMapping(path = "/deleteHidePost", params="action=delete")
     public String deletePost(@RequestParam("idPost") Integer idPost) {
-        //Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        logger.info("idPost: " + idPost);
-    	postService.deletePostByIdPost(idPost);
-    	//deleting orphans tags that do not belong to any post
-    	tagService.deleteOrphans();
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        User loggedUser = userService.findByUsername(auth.getName());
+        Post postToDelete = postService.findByIdPost(idPost);
+        //check that the user have permissions to delete
+        if(loggedUser.getUsername().equals(postToDelete.getUser().getUsername())) {
+        	postService.deletePostByIdPost(idPost);
+        	//deleting orphans tags that do not belong to any post
+        	tagService.deleteOrphans();
+        }
+        else logger.info("user: " + auth.getName() + " tried to delete post " + idPost + " but does not have permissions!");
+        
     	return "redirect:/posts";
     }
     
-    @PostMapping(path = "/posts")
+    @PostMapping(path = "/deleteHidePost", params="action=hide")
+    public String hidePost(@RequestParam("idPost") Integer idPost) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        User loggedUser = userService.findByUsername(auth.getName());
+        Post postToHide = postService.findByIdPost(idPost);
+        //check that the user have permissions to delete
+        if (loggedUser.getUserType().equals("ADMIN_ROLE") || loggedUser.getUsername().equals(postToHide.getUser().getUsername())) {
+        	postService.hidePost(idPost, loggedUser.getIdUser());
+        }
+        else logger.info("user: " + auth.getName() + " tried to hide post " + idPost + " but does not have permissions!");
+        
+    	return "redirect:/posts";
+    }
+    
+    @PostMapping(path = "/posts", params="action=search")
     public ModelAndView seatchPost(@RequestParam("idUser") Integer idUser, @RequestParam("text") String text, @RequestParam("idTag") Integer idTag) {
     	ModelAndView posts = new ModelAndView("posts");
-    	//Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        logger.info("idUser: " + idUser + "text: " + text + "tagId :" + idTag);
+    	Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        logger.info("idUser: " + idUser + " ,text: " + text + " ,tagId :" + idTag);
         posts.addObject("postList", postService.searchPosts(idUser, text, idTag));
-    	posts.addObject("post", new Post());
+    	posts.addObject("post", new Post(userService.findByUsername(auth.getName())));
     	posts.addObject("userList", userService.findAll()); 
     	posts.addObject("tagList", tagService.findAll());
+    	posts.addObject("loggedUser", userService.findByUsername(auth.getName()));
     	
         //return "redirect:/posts";
     	return posts;
